+++ packages/core/babel-preset-dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # babel-preset-dlight
 
+## 1.0.0-beta.13
+
+### Patch Changes
+
+- fix: no updateView in static property/method/non-updateDerivedBlock
+
 ## 1.0.0-beta.12
 
 ### Patch Changes